added vs code launch template suppports
